backend/requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.24.0
sqlmodel==0.0.31
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.2.1
better-auth==0.0.1b11
psycopg2==2.9.11
python-jose==3.5.0
cryptography==46.0.4
